+++ bonus/exec_child_bonus.c
@@ -6,7 +6,7 @@
 /*   By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/18 22:38:25 by irychkov          #+#    #+#             */
-/*   Updated: 2024/09/21 13:10:58 by irychkov         ###   ########.fr       */
+/*   Updated: 2024/09/21 14:23:14 by irychkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static void	fd_out_init(t_pipex *fds, char *av[])
 {
-	int file_index;
+	int	file_index;
 
 	if (fds->here_doc)
 		file_index = fds->num_cmds + 3;
@@ -26,42 +26,23 @@ static void	fd_out_init(t_pipex *fds, char *av[])
 	if (access(av[file_index], F_OK) == 0 && access(av[file_index], W_OK) == -1)
 		error_permission(av[file_index], 1, fds);
 	if (fds->here_doc)
-		fds->pipex[1] = open(av[file_index], O_WRONLY | O_CREAT | O_APPEND, 0644);
+		fds->pipex[1] = open(av[file_index], O_WRONLY
+				| O_CREAT | O_APPEND, 0644);
 	else
-		fds->pipex[1] = open(av[file_index], O_WRONLY | O_CREAT | O_TRUNC, 0644);
+		fds->pipex[1] = open(av[file_index], O_WRONLY
+				| O_CREAT | O_TRUNC, 0644);
 	if (fds->pipex[1] < 0)
 		error_nofile(av[file_index], 1, fds);
 }
 
-void	fd_in_init(t_pipex *fds, char *av[])
-{
-	if (access(av[1], F_OK) == 0 && access(av[1], R_OK) == -1)
-		error_permission(av[1], 126, fds);
-	else
-	{
-		fds->pipex[0] = open(av[1], O_RDONLY);
-		if (fds->pipex[0] < 0)
-			error_nofile(av[1], 127, fds);
-	}
-}
-
-void	redirect_error(t_pipex *fds, int cmd_pos)
+static void	redirect_error(t_pipex *fds, int cmd_pos)
 {
 	if (dup2(fds->error_fds[cmd_pos], STDERR_FILENO) == -1)
 		error_dup(fds);
 	close(fds->error_fds[cmd_pos]);
 }
 
-static void	execute_command(char *cmd, char **envp, t_pipex *fds)
-{
-	if (!cmd || cmd[0] == '\0')
-		error_permission(cmd, 126, fds);
-	if (cmd[0] == ' ')
-		error_command(cmd, fds);
-	exec_with_zsh(cmd, envp, fds);
-}
-
-void	read_from_pipe(t_pipex *fds, int cmd_pos)
+static void	read_from_pipe(t_pipex *fds, int cmd_pos)
 {
 	if (cmd_pos == 0)
 	{
@@ -78,7 +59,7 @@ void	read_from_pipe(t_pipex *fds, int cmd_pos)
 	}
 }
 
-void	write_to_pipe(t_pipex *fds, int cmd_pos, char *av[])
+static void	write_to_pipe(t_pipex *fds, int cmd_pos, char *av[])
 {
 	if (cmd_pos == fds->num_cmds - 1)
 	{
@@ -98,7 +79,7 @@ void	write_to_pipe(t_pipex *fds, int cmd_pos, char *av[])
 
 void	exec_child(int cmd_pos, t_pipex *fds, char *av[], char **envp)
 {
-	int cmd_index;
+	int	cmd_index;
 
 	if (fds->here_doc)
 		cmd_index = cmd_pos + 3;
