+++ bonus/free_bonus.c
@@ -6,7 +6,7 @@
 /*   By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/18 21:07:02 by irychkov          #+#    #+#             */
-/*   Updated: 2024/09/19 00:21:31 by irychkov         ###   ########.fr       */
+/*   Updated: 2024/09/21 14:28:32 by irychkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,31 +19,25 @@ void	close_pipes(t_pipex *fds)
 	i = 0;
 	while (i < fds->num_cmds - 1)
 	{
-		if(fds->pipes[i][0] >= 0)
+		if (fds->pipes[i][0] >= 0)
 			close(fds->pipes[i][0]);
-		if(fds->pipes[i][1] >= 0)
+		if (fds->pipes[i][1] >= 0)
 			close(fds->pipes[i][1]);
 		free(fds->pipes[i]);
 		i++;
 	}
 	free(fds->pipes);
-	
 	i = 0;
 	while (i < fds->num_cmds)
 	{
-		if(fds->error_fds[i] >= 0)
+		if (fds->error_fds[i] >= 0)
 			close(fds->error_fds[i]);
 		i++;
 	}
 	free(fds->error_fds);
-
-	if(fds->fd[0] >= 0)
-		close(fds->fd[0]);
-	if(fds->fd[1] >= 0)
-		close(fds->fd[1]);
-	if(fds->pipex[0] >= 0)
+	if (fds->pipex[0] >= 0)
 		close(fds->pipex[0]);
-	if(fds->pipex[1] >= 0)
+	if (fds->pipex[1] >= 0)
 		close(fds->pipex[1]);
 }
